// packages/generator/src/emitter/dmmf/types.ts
export type ReadonlyDeep<O> = {
  +readonly [K in keyof O]: ReadonlyDeep<O[K]>;
};

export namespace DMMF {
  export type Document = ReadonlyDeep<{
    datamodel: Datamodel;
    schema: Schema;
    // mappings: Mappings;
    // additional props:
    modelMappings: ModelMapping[];
  }>;
  export type Enum = ReadonlyDeep<{
    name: string;
    // values: EnumValue[];
    dbName?: string | null;
    // documentation?: string;
    // additional props:
    typeName: string;
    docs: string | undefined;
    valuesMap: Array<{ name: string; value: string }>;
  }>;
  export type Datamodel = ReadonlyDeep<{
    models: Model[];
    enums: Enum[];
    types: Model[];
  }>;
// ...
  export type Schema = ReadonlyDeep<{
    rootQueryType?: string;
    rootMutationType?: string;
    inputTypes: InputType[];
    outputTypes: OutputType[];
    enums: Enum[];
    // fieldRefTypes: {
    //   prisma?: FieldRefType[];
    // };
  }>;
// ...
  export type OutputType = ReadonlyDeep<{
    name: string;
    // fields: SchemaField[];
    // additional props:
    fields: OutputSchemaField[];
    typeName: string;
  }>;
// ...
  export type OutputSchemaField = SchemaField &
    ReadonlyDeep<{
      argsTypeName: string | undefined;
    }>;
// ...
  export type ModelMapping = ReadonlyDeep<{
    // model: string;
    // plural: string;
    // findUnique?: string | null;
    // findUniqueOrThrow?: string | null;
    // findFirst?: string | null;
    // findFirstOrThrow?: string | null;
    // findMany?: string | null;
    // createOne?: string | null;
    // createMany?: string | null;
    // createManyAndReturn?: string | null;
    // updateOne?: string | null;
    // updateMany?: string | null;
    // upsertOne?: string | null;
    // deleteOne?: string | null;
    // deleteMany?: string | null;
    // aggregate?: string | null;
    // groupBy?: string | null;
    // count?: string | null;
    // findRaw?: string | null;
    // aggregateRaw?: string | null;

    // additional props
    modelName: string;
    actions: Action[];
    collectionName: string;
    resolverName: string;
    modelTypeName: string;
  }>;
// ...
}